@@ -30,17 +30,17 @@
     "vuetify": "^3.7.4"
   },
   "devDependencies": {
-    "@electron-toolkit/eslint-config": "^1.0.2",
+    "@electron-toolkit/eslint-config": "^2.0.0",
     "@rushstack/eslint-patch": "^1.10.3",
     "@vitejs/plugin-vue": "^5.0.5",
-    "@vue/eslint-config-prettier": "^9.0.0",
+    "@vue/eslint-config-prettier": "^10.2.0",
     "archiver": "^7.0.1",
     "cheerio": "^1.0.0",
     "electron": "^31.0.2",
     "electron-builder": "^24.13.3",
     "electron-vite": "^2.3.0",
-    "eslint": "^8.57.0",
-    "eslint-plugin-vue": "^9.26.0",
+    "eslint": "^9.21.0",
+    "eslint-plugin-vue": "^9.32.0",
     "pinia": "^2.2.6",
     "prettier": "^3.3.2",
     "vite": "^5.3.1",
