@@ -5,7 +5,7 @@ const path = require('path')
 // Deletes unreferenced projects and images. This cannot happen immediately
 // upon user action to provide undo functionality.
 
-function cleanProjects(vianPath, projects) {
+const cleanProjects = (vianPath, projects) => {
   fs.readdirSync(vianPath, { withFileTypes: true })
     .filter((i) => i.isDirectory())
     .forEach((d) => {
@@ -16,25 +16,21 @@ function cleanProjects(vianPath, projects) {
     })
 }
 
-function cleanScreenshots(projectPath) {
+const cleanScreenshots = (projectPath) => {
   const content = fs.readFileSync(path.join(projectPath, 'undoable.json'), 'utf8')
-  const timelines = JSON.parse(content).timelines
+  const { timelines } = JSON.parse(content)
   const referencedImgs = new Set()
   timelines.forEach((t) => {
     t.data.forEach((d) => {
-      if (d.image !== undefined) referencedImgs.add(d.image.replace('app://', ''))
-      if (d.thumbnail !== undefined) referencedImgs.add(d.thumbnail.replace('app://', ''))
+      if ('image' in d) referencedImgs.add(d.image.replace('app://', ''))
+      if ('thumbnail' in d) referencedImgs.add(d.thumbnail.replace('app://', ''))
     })
   })
 
-  let fsImgs
-  try {
-    fsImgs = fs
-      .readdirSync(path.join(projectPath, 'screenshots'), { withFileTypes: true })
-      .map((i) => path.join(i.parentPath, i.name))
-  } catch (e) {
-    return
-  }
+  if (!fs.existsSync(path.join(projectPath, 'screenshots'))) return
+  const fsImgs = fs
+    .readdirSync(path.join(projectPath, 'screenshots'), { withFileTypes: true })
+    .map((i) => path.join(i.parentPath, i.name))
 
   const diff = new Set(fsImgs).difference(referencedImgs)
   diff.forEach((i) => {
@@ -43,15 +39,17 @@ function cleanScreenshots(projectPath) {
   if (diff.size > 0) console.log('Deleted %d screenshots', diff.size)
 }
 
-if (workerData !== null && workerData !== undefined) {
-  console.log('Started cleanup worker')
-  const vianPath = workerData
-  const content = fs.readFileSync(path.join(vianPath, 'meta.json'), 'utf8')
-  const projects = new Set(JSON.parse(content).projects.map((p) => p.id))
+console.log('Started cleanup worker')
+const vianPath = workerData
+const content = fs.readFileSync(path.join(vianPath, 'meta.json'), 'utf8')
+const projects = new Set(JSON.parse(content).projects.map((p) => p.id))
 
-  cleanProjects(vianPath, projects)
-  projects.forEach((p) => {
+cleanProjects(vianPath, projects)
+projects.forEach((p) => {
+  try {
     cleanScreenshots(path.join(vianPath, p))
-  })
-  console.log('Cleanup done')
-}
+  } catch (e) {
+    console.log('Error cleaning screenshots:', e)
+  }
+})
+console.log('Cleanup done')
