@@ -4,7 +4,7 @@ const path = require('path')
 const os = require('os')
 const archiver = require('archiver')
 
-async function exportScreenshots(storePath, location, frames) {
+const exportScreenshots = async (storePath, location, frames) => {
   const tmpPath = fs.mkdtempSync(path.join(os.tmpdir(), 'vian-screenshots-'))
 
   const store = JSON.parse(fs.readFileSync(storePath, 'utf8'))
@@ -15,13 +15,13 @@ async function exportScreenshots(storePath, location, frames) {
     fs.mkdirSync(timelinePath)
     t.data.forEach((s) => {
       const imagePath = s.image.replace('app://', '')
-      if (frames && !frames.includes(Number(path.basename(imagePath).replace('.jpg', '')))) return
+      if (frames && !frames.includes(s.frame)) return
       fs.copyFileSync(imagePath, path.join(timelinePath, path.basename(imagePath)))
     })
   })
 
-  if (!location.endsWith('.zip')) location += '.zip'
-  const output = fs.createWriteStream(location)
+  const finalLocation = location.endsWith('.zip') ? location : `${location}.zip`
+  const output = fs.createWriteStream(finalLocation)
   const archive = archiver('zip')
 
   archive.pipe(output)
@@ -31,9 +31,8 @@ async function exportScreenshots(storePath, location, frames) {
 }
 
 console.log('Started screenshot export worker')
-if (workerData !== null && workerData !== undefined) {
-  exportScreenshots(workerData.storePath, workerData.location, workerData.frames).then((err) => {
-    if (err) throw err
-    else parentPort.postMessage(true)
-  })
-}
+
+exportScreenshots(workerData.storePath, workerData.location, workerData.frames).then((err) => {
+  if (err) throw err
+  else parentPort.postMessage(true)
+})
