@@ -5,13 +5,15 @@
 </template>
 
 <script>
+import api from '@renderer/api'
 import { useMainStore } from '@renderer/stores/main'
 import { useMetaStore } from '@renderer/stores/meta'
-import { useUndoableStore } from '@renderer/stores/undoable'
 import { useTempStore } from '@renderer/stores/temp'
-import api from '@renderer/api'
+import { useUndoableStore } from '@renderer/stores/undoable'
 
 export default {
+  name: 'App',
+
   created() {
     useMainStore().initialize()
     useUndoableStore().initialize()
