@@ -2,29 +2,36 @@
   <v-app-bar density="compact">
     <v-app-bar-title
       >VIAN-lite
-      <span class="text-medium-emphasis text-body-2">{{ mainStore.video }}</span></v-app-bar-title
+      <span class="text-body-2 text-medium-emphasis">{{ mainStore.video }}</span></v-app-bar-title
     >
+
     <v-btn :disabled="!isUndoable" icon @click="undo">
       <v-icon>mdi-undo</v-icon>
     </v-btn>
+
     <v-btn :disabled="!isRedoable" icon @click="redo">
       <v-icon>mdi-redo</v-icon>
     </v-btn>
+
     <v-menu :close-on-content-click="false">
       <template #activator="{ props }">
         <v-btn v-tooltip="'Job list'" :disabled="!hasJobs" icon v-bind="props">
           <v-badge v-if="runningJobs" color="error" dot>
             <v-icon>mdi-format-list-bulleted</v-icon>
           </v-badge>
+
           <v-icon v-else>mdi-format-list-bulleted</v-icon>
         </v-btn>
       </template>
+
       <v-list>
         <v-list-item v-for="(job, i) in tempStore.jobs" :key="i">
           <v-list-item-title>{{ job.type }}</v-list-item-title>
+
           <template #append>
             <v-badge :color="statusToColor(job.status)" :content="job.status" inline></v-badge>
           </template>
+
           <template #prepend>
             <v-btn
               v-if="job.status === 'RUNNING'"
@@ -46,48 +53,65 @@
           <v-icon>mdi-tools</v-icon>
         </v-btn>
       </template>
+
       <v-list>
         <v-list-item @click="shotBoundaryDetectionClicked">
           <v-list-item-title>Shotboundary Detection</v-list-item-title>
         </v-list-item>
+
         <v-list-item @click="loadSubtitles">
           <v-list-item-title>Load Subtitles</v-list-item-title>
         </v-list-item>
+
         <v-list-item @click="genScreenshotDialog = true">
           <v-list-item-title>Generate Screenshots</v-list-item-title>
         </v-list-item>
+
         <v-list-item @click="importAnnotations">
           <v-list-item-title>Import Annotations</v-list-item-title>
         </v-list-item>
+
         <v-list-item>
           <v-list-item-title>Export</v-list-item-title>
+
           <template #append>
             <v-icon icon="mdi-menu-right" size="x-small"></v-icon>
           </template>
+
           <v-menu :open-on-focus="false" activator="parent" open-on-hover submenu>
             <v-list>
               <v-list-item @click="exportAnnotations(false)">
                 <v-list-item-title>Export Annotations (eaf)</v-list-item-title>
               </v-list-item>
+
               <v-list-item @click="exportAnnotations(true)">
                 <v-list-item-title>Export Annotations (csv)</v-list-item-title>
               </v-list-item>
+
               <v-list-item @click="exportScreenshotsDialog = true">
                 <v-list-item-title>Export Screenshots</v-list-item-title>
+              </v-list-item>
+
+              <v-list-item @click="exportProject">
+                <v-list-item-title>Export Project</v-list-item-title>
               </v-list-item>
             </v-list>
           </v-menu>
         </v-list-item>
+
         <v-list-item>
           <v-list-item-title>Layout</v-list-item-title>
+
           <template #append>
             <v-icon icon="mdi-menu-right" size="x-small"></v-icon>
           </template>
+
           <v-menu :open-on-focus="false" activator="parent" open-on-hover submenu>
             <v-list>
               <v-list-item prepend-icon="mdi-tab" @click="layout = 'tibava'">
                 <v-list-item-title>Tabbed</v-list-item-title>
               </v-list-item>
+
               <v-list-item prepend-icon="mdi-arrow-all" @click="layout = 'draggable'">
                 <v-list-item-title>Draggable</v-list-item-title>
               </v-list-item>
@@ -96,16 +120,21 @@
         </v-list-item>
       </v-list>
     </v-menu>
+
     <v-btn v-tooltip="'Home'" icon @click="homeClicked">
       <v-icon>mdi-home</v-icon>
     </v-btn>
   </v-app-bar>
+
   <v-main class="ma-3">
-    <layout-tibava v-if="layout === 'tibava'"></layout-tibava>
-    <layout-draggable v-else-if="layout === 'draggable'"></layout-draggable>
+    <LayoutTibava v-if="layout === 'tibava'"></LayoutTibava>
+
+    <LayoutDraggable v-else-if="layout === 'draggable'"></LayoutDraggable>
+
     <v-dialog v-model="genScreenshotDialog" persistent max-width="400">
       <v-card>
         <v-card-title>Generate Screenshots</v-card-title>
+
         <v-card-text>
           Create a screenshot every N seconds
           <v-text-field
@@ -115,10 +144,12 @@
             min="1"
             max="100"
           ></v-text-field>
+
           <v-checkbox
             v-model="screenshotPerShot"
             label="Ensure at least one screenshot per shot"
           ></v-checkbox>
+
           <v-select
             v-if="screenshotPerShot"
             v-model="screenshotShotTimeline"
@@ -129,8 +160,10 @@
           >
           </v-select>
         </v-card-text>
+
         <v-card-actions>
           <v-btn color="warning" @click="genScreenshotDialog = false">Cancel</v-btn>
+
           <v-btn
             color="primary"
             :disabled="genScreenshotButtonDisabled"
@@ -140,16 +173,21 @@
         </v-card-actions>
       </v-card>
     </v-dialog>
+
     <v-dialog v-model="exportScreenshotsDialog" persistent max-width="500">
       <v-card>
         <v-card-title>Export Screenshots</v-card-title>
+
         <v-card-text>
           You can either export all screenshots at once or select them individually in a timeline
           and only export those.
         </v-card-text>
+
         <v-card-actions>
           <v-btn color="warning" @click="exportScreenshotsDialog = false">Cancel</v-btn>
+
           <v-btn color="primary" @click="exportScreenshots(false)">Export all</v-btn>
+
           <v-btn
             color="secondary"
             :disabled="exportScreenshotsIndividualDisabled"
@@ -165,73 +203,73 @@
 <script>
 import { mapStores } from 'pinia'
 
+import { exportAnnotations, exportProject } from '@renderer/importexport'
+import LayoutDraggable from '@renderer/components/LayoutDraggable.vue'
+import LayoutTibava from '@renderer/components/LayoutTibava.vue'
+import api from '@renderer/api'
 import { useMainStore } from '@renderer/stores/main'
 import { useTempStore } from '@renderer/stores/temp'
-import { useUndoableStore } from '@renderer/stores/undoable'
 import { useUndoStore } from '@renderer/stores/undo'
-import api from '@renderer/api'
-import { exportAnnotations } from '@renderer/importexport'
-import LayoutTibava from '@renderer/components/LayoutTibava.vue'
-import LayoutDraggable from '@renderer/components/LayoutDraggable.vue'
+import { useUndoableStore } from '@renderer/stores/undoable'
 
 export default {
-  components: { LayoutTibava, LayoutDraggable },
+  name: 'Id',
+  components: { LayoutDraggable, LayoutTibava },
+
   data: () => ({
-    layout: 'tibava',
+    exportScreenshotsDialog: false,
     genScreenshotDialog: false,
-    screenshotPerShot: false,
+    layout: 'tibava',
     screenshotInterval: 10,
-    screenshotShotTimeline: undefined,
-    exportScreenshotsDialog: false
+    screenshotPerShot: false,
+    screenshotShotTimeline: null
   }),
+
   computed: {
     ...mapStores(useMainStore, useTempStore, useUndoableStore, useUndoStore),
-    runningJobs() {
-      return this.tempStore.jobs.some((j) => j.status === 'RUNNING')
+
+    exportScreenshotsIndividualDisabled() {
+      if (this.tempStore.selectedSegments.size === 0) return true
+      const [shotid, timelineid] = this.tempStore.selectedSegments.entries().next().value
+      const segment = this.undoableStore.getSegmentForId(timelineid, shotid)
+      return 'image' in segment
     },
+
+    genScreenshotButtonDisabled() {
+      return this.screenshotPerShot && this.screenshotShotTimeline === null
+    },
+
     hasJobs() {
       return this.tempStore.jobs.length > 0
     },
-    isUndoable() {
-      return this.undoStore.isUndoable('undoable')
-    },
+
     isRedoable() {
       return this.undoStore.isRedoable('undoable')
     },
-    genScreenshotButtonDisabled() {
-      return this.screenshotPerShot && this.screenshotShotTimeline === undefined
+
+    isUndoable() {
+      return this.undoStore.isUndoable('undoable')
     },
-    exportScreenshotsIndividualDisabled() {
-      if (this.tempStore.selectedSegments.size == 0) return true
-      const [shotid, timelineid] = this.tempStore.selectedSegments.entries().next().value
-      const segment = this.undoableStore.getSegmentForId(timelineid, shotid)
-      return segment.image === undefined
+
+    runningJobs() {
+      return this.tempStore.jobs.some((j) => j.status === 'RUNNING')
     }
   },
+
   created() {
     this.mainStore.loadStore(this.$route.params.id)
     this.undoableStore.loadStore(this.$route.params.id)
   },
+
   methods: {
-    homeClicked() {
-      this.$router.push('/')
-      this.mainStore.$reset()
-      this.tempStore.$reset()
-      this.undoableStore.$reset()
-      this.undoStore.reset()
+    exportAnnotations(csv) {
+      exportAnnotations(csv)
     },
-    undo() {
-      this.undoableStore.undo('undoable')
+
+    exportProject() {
+      exportProject()
     },
-    redo() {
-      this.undoableStore.redo('undoable')
-    },
-    shotBoundaryDetectionClicked() {
-      this.undoableStore.runShotBoundaryDetection()
-    },
-    loadSubtitles() {
-      this.undoableStore.loadSubtitles()
-    },
+
     exportScreenshots(individually) {
       if (individually) {
         const frames = Array.from(this.tempStore.selectedSegments).map(
@@ -243,16 +281,17 @@ export default {
       }
       this.exportScreenshotsDialog = false
     },
+
     generateScreenshots() {
       this.genScreenshotDialog = false
-      let frames = []
+      const frames = []
       let shotI = 0
       const timelineShots = this.screenshotPerShot
         ? this.undoableStore.shotTimelines.find((t) => t.id === this.screenshotShotTimeline).data
         : []
 
-      for (let i = 0; i <= this.mainStore.videoDuration; i++) {
-        if (i % this.screenshotInterval == 0) {
+      for (let i = 0; i <= this.mainStore.videoDuration; i += 1) {
+        if (i % this.screenshotInterval === 0) {
           const frame = Math.floor(i * this.mainStore.fps)
 
           while (
@@ -268,25 +307,48 @@ export default {
                 )
               )
             }
-            shotI++
+            shotI += 1
           }
           frames.push(frame)
         }
       }
       this.undoableStore.generateScreenshots(frames)
     },
+
+    homeClicked() {
+      this.$router.push('/')
+      this.mainStore.$reset()
+      this.tempStore.$reset()
+      this.undoableStore.$reset()
+      this.undoStore.reset()
+    },
+
     importAnnotations() {
       this.undoableStore.importAnnotations()
     },
-    exportAnnotations(csv) {
-      exportAnnotations(csv)
+
+    loadSubtitles() {
+      this.undoableStore.loadSubtitles()
     },
+
+    redo() {
+      this.undoableStore.redo('undoable')
+    },
+
+    shotBoundaryDetectionClicked() {
+      this.undoableStore.runShotBoundaryDetection()
+    },
+
     statusToColor(status) {
       if (status === 'RUNNING') return 'yellow'
       else if (status === 'ERROR') return 'error'
       else if (status === 'DONE') return 'success'
       else if (status === 'CANCELED') return 'black'
-      else return 'white'
+      return 'white'
+    },
+
+    undo() {
+      this.undoableStore.undo('undoable')
     }
   }
 }
