@@ -4,19 +4,17 @@ export const useUndoStore = defineStore('undo', {
   state: () => ({
     stack: {}
   }),
+  /* eslint-disable-next-line vue/sort-keys */
   actions: {
-    isUndoable(name) {
-      return this.stack[name] !== undefined && this.stack[name].index > 0
-    },
     isRedoable(name) {
-      return (
-        this.stack[name] !== undefined &&
-        this.stack[name].index < this.stack[name].entries.length - 1
-      )
+      return name in this.stack && this.stack[name].index < this.stack[name].entries.length - 1
+    },
+    isUndoable(name) {
+      return name in this.stack && this.stack[name].index > 0
     },
     push(name, store) {
-      if (this.stack[name] === undefined) {
-        this.stack[name] = { index: -1, entries: [], ignoreNext: false }
+      if (!(name in this.stack)) {
+        this.stack[name] = { entries: [], ignoreNext: false, index: -1 }
       }
       if (this.stack[name].ignoreNext) {
         this.stack[name].ignoreNext = false
@@ -25,25 +23,19 @@ export const useUndoStore = defineStore('undo', {
 
       this.stack[name].entries = this.stack[name].entries.slice(0, this.stack[name].index + 1)
       this.stack[name].entries.push(store)
-      this.stack[name].index++
+      this.stack[name].index += 1
       if (this.stack[name].entries.length > 30) {
         this.stack[name].entries.shift()
-        this.stack[name].index--
-      }
-    },
-    undo(name) {
-      if (this.isUndoable(name)) {
-        this.stack[name].index--
-        this.stack[name].ignoreNext = true
-        return JSON.parse(JSON.stringify(this.stack[name].entries[this.stack[name].index]))
+        this.stack[name].index -= 1
       }
     },
     redo(name) {
       if (this.isRedoable(name)) {
-        this.stack[name].index++
+        this.stack[name].index += 1
         this.stack[name].ignoreNext = true
         return JSON.parse(JSON.stringify(this.stack[name].entries[this.stack[name].index]))
       }
+      throw new Error('Not redoable')
     },
     reset() {
       for (const key of Object.keys(this.stack)) {
@@ -51,6 +43,14 @@ export const useUndoStore = defineStore('undo', {
           delete this.stack[key]
         }
       }
+    },
+    undo(name) {
+      if (this.isUndoable(name)) {
+        this.stack[name].index -= 1
+        this.stack[name].ignoreNext = true
+        return JSON.parse(JSON.stringify(this.stack[name].entries[this.stack[name].index]))
+      }
+      throw new Error('Not undoable')
     }
   }
 })
