@@ -1,135 +1,59 @@
 import { defineStore } from 'pinia'
 
-import { useUndoStore } from './undo'
-import { useMainStore } from './main'
-import { useTempStore } from './temp'
 import api from '@renderer/api'
 import { parseEafAnnotations } from '@renderer/importexport'
+import { useMainStore } from './main'
+import { useTempStore } from './temp'
+import { useUndoStore } from './undo'
 
 export const useUndoableStore = defineStore('undoable', {
   state: () => ({
+    id: null,
     subtitles: null,
-    timelines: [],
     subtitlesVisible: true,
-    id: null
+    timelines: []
   }),
+  /* eslint-disable-next-line vue/sort-keys */
   getters: {
-    shotTimelines: (state) => state.timelines.filter((t) => t.type === 'shots'),
-    screenshotTimelines: (state) => state.timelines.filter((t) => t.type.startsWith('screenshot'))
+    screenshotTimelines: (state) => state.timelines.filter((t) => t.type.startsWith('screenshot')),
+    shotTimelines: (state) => state.timelines.filter((t) => t.type === 'shots')
   },
+  /* eslint-disable-next-line vue/sort-keys */
   actions: {
-    async initialize() {
-      this.$subscribe((mutation, state) => {
-        if (state.id === null || mutation.events.key === 'id') return
-        const copyState = JSON.parse(JSON.stringify(state))
-        api.saveStore('undoable', copyState)
-
-        useUndoStore().push('undoable', copyState)
-      })
-
-      api.onScreenshotsGeneration(this.onScreenshotsGeneration)
-      api.onScreenshotGeneration(this.onScreenshotGeneration)
-      api.onShotBoundaryDetection(this.onShotBoundaryDetection)
-    },
-    onShotBoundaryDetection(data) {
+    addNewTimeline() {
       this.timelines.push({
-        type: 'shots',
+        data: [],
+        id: crypto.randomUUID(),
         name: 'Shots',
-        id: crypto.randomUUID(),
-        data: data.map((shot) => ({ start: shot[0], end: shot[1], id: crypto.randomUUID() }))
+        type: 'shots'
       })
-    },
-    onScreenshotGeneration(data) {
-      data.id = crypto.randomUUID()
-      if (this.timelines.find((t) => t.type === 'screenshots-manual')) {
-        this.timelines.find((t) => t.type === 'screenshots-manual').data.push(data)
-      } else {
-        this.timelines.push({
-          type: 'screenshots-manual',
-          name: 'Manual Screenshots',
-          id: crypto.randomUUID(),
-          data: [data]
-        })
-      }
-    },
-    onScreenshotsGeneration(data) {
-      data.forEach((screenshot) => {
-        screenshot.id = crypto.randomUUID()
-      })
-      this.timelines.push({
-        type: 'screenshots',
-        name: 'Screenshots',
-        id: crypto.randomUUID(),
-        data: data
-      })
-    },
-    async runShotBoundaryDetection() {
-      api.runShotBoundaryDetection(useMainStore().video)
-    },
-    async loadSubtitles() {
-      this.subtitles = await api.loadSubtitles(this.id)
-      return this.subtitles
-    },
-    getTimelineById(id) {
-      return this.timelines.find((t) => t.id === id)
-    },
-    deleteSegments(timelineId, segmentIds) {
-      const timeline = this.getTimelineById(timelineId)
-      timeline.data = timeline.data.filter((s) => !segmentIds.includes(s.id))
-      useTempStore().validateSelectedSegments()
-    },
-    mergeSegments(timelineId, segmentIds) {
-      const timeline = this.getTimelineById(timelineId)
-      const segment = timeline.data.find((s) => s.id == segmentIds[0])
-      segmentIds.slice(1).forEach((segmentId) => {
-        const index = timeline.data.findIndex((s) => s.id === segmentId)
-        segment.start = Math.min(segment.start, timeline.data[index].start)
-        segment.end = Math.max(segment.end, timeline.data[index].end)
-        timeline.data.splice(index, 1)
-      })
-      useTempStore().validateSelectedSegments()
-    },
-    getSegmentForId(timelineId, segmentId) {
-      const timeline = this.getTimelineById(timelineId)
-      return timeline.data.find((s) => s.id === segmentId)
-    },
-    splitSegment(timelineId, segmentId, position) {
-      const timeline = this.getTimelineById(timelineId)
-      const index = timeline.data.findIndex((s) => s.id === segmentId)
-      const segment = timeline.data[index]
-      timeline.data.splice(index + 1, 0, {
-        start: position,
-        end: segment.end,
-        id: crypto.randomUUID()
-      })
-      segment.end = position - 1
     },
     addShotToNth(n, start, end) {
       const data = this.timelines[n].data.slice()
       data.push({
-        start: Math.round(start),
         end: Math.round(end),
-        id: crypto.randomUUID()
+        id: crypto.randomUUID(),
+        start: Math.round(start)
       })
       data.sort((a, b) => a.start - b.start)
       this.timelines[n].data = data
     },
     changeShotBoundaries(shotId, start, end) {
       for (const timeline of this.timelines) {
-        if (timeline.type !== 'shots') continue
-        const shot = timeline.data.find((s) => s.id === shotId)
-        if (shot) {
-          shot.start = start
-          shot.end = end
-          return
+        if (timeline.type === 'shots') {
+          const shot = timeline.data.find((s) => s.id === shotId)
+          if (shot) {
+            shot.start = start
+            shot.end = end
+            return
+          }
         }
       }
     },
-    generateScreenshots(frames) {
-      api.runScreenshotsGeneration(useMainStore().video, frames, this.id)
-    },
-    generateScreenshot(frame) {
-      api.runScreenshotGeneration(useMainStore().video, frame, this.id)
+    deleteSegments(timelineId, segmentIds) {
+      const timeline = this.getTimelineById(timelineId)
+      timeline.data = timeline.data.filter((s) => !segmentIds.includes(s.id))
+      useTempStore().validateSelectedSegments()
     },
     deleteTimeline(id) {
       this.timelines = this.timelines.filter((t) => t.id !== id)
@@ -145,27 +69,29 @@ export const useUndoableStore = defineStore('undoable', {
       newTimeline.name += ' (copy)'
       this.timelines.push(newTimeline)
     },
-    renameTimeline(id, name) {
-      this.getTimelineById(id).name = name
+    generateScreenshot(frame) {
+      api.runScreenshotGeneration(useMainStore().video, frame, this.id)
     },
-    addNewTimeline() {
-      this.timelines.push({
-        type: 'shots',
-        name: 'Shots',
-        id: crypto.randomUUID(),
-        data: []
-      })
+    generateScreenshots(frames) {
+      api.runScreenshotsGeneration(useMainStore().video, frames, this.id)
+    },
+    getSegmentForId(timelineId, segmentId) {
+      const timeline = this.getTimelineById(timelineId)
+      return timeline.data.find((s) => s.id === segmentId)
+    },
+    getTimelineById(id) {
+      return this.timelines.find((t) => t.id === id)
     },
     importAnnotations() {
       const fileInput = document.createElement('input')
       fileInput.type = 'file'
       fileInput.accept = '.eaf'
-      fileInput.addEventListener('change', async (event) => {
-        const file = event.target.files[0]
+      fileInput.addEventListener('change', (event) => {
+        const [file] = event.target.files
         if (!file) return
 
         const reader = new FileReader()
-        reader.onload = async (e) => {
+        reader.onload = (e) => {
           const content = e.target.result
           this.timelines = this.timelines.concat(parseEafAnnotations(content))
         }
@@ -173,19 +99,96 @@ export const useUndoableStore = defineStore('undoable', {
       })
       fileInput.click()
     },
+    initialize() {
+      this.$subscribe((mutation, state) => {
+        if (state.id === null || mutation.events?.key === 'id') return
+        const copyState = JSON.parse(JSON.stringify(state))
+        api.saveStore('undoable', copyState)
+
+        useUndoStore().push('undoable', copyState)
+      })
+
+      api.onScreenshotsGeneration(this.onScreenshotsGeneration)
+      api.onScreenshotGeneration(this.onScreenshotGeneration)
+      api.onShotBoundaryDetection(this.onShotBoundaryDetection)
+    },
     async loadStore(projectId) {
       const state = await api.loadStore('undoable', projectId)
-      if (state !== null) {
-        this.$patch(state)
-      } else {
+      if (state === null) {
         this.id = projectId
+      } else {
+        this.$patch(state)
       }
     },
-    undo() {
-      this.$patch(useUndoStore().undo('undoable'))
+    async loadSubtitles() {
+      this.subtitles = await api.loadSubtitles(this.id)
+      return this.subtitles
+    },
+    mergeSegments(timelineId, segmentIds) {
+      const timeline = this.getTimelineById(timelineId)
+      const segment = timeline.data.find((s) => s.id === segmentIds[0])
+      segmentIds.slice(1).forEach((segmentId) => {
+        const index = timeline.data.findIndex((s) => s.id === segmentId)
+        segment.start = Math.min(segment.start, timeline.data[index].start)
+        segment.end = Math.max(segment.end, timeline.data[index].end)
+        timeline.data.splice(index, 1)
+      })
+      useTempStore().validateSelectedSegments()
+    },
+    onScreenshotGeneration(data) {
+      data.id = crypto.randomUUID()
+      if (this.timelines.find((t) => t.type === 'screenshots-manual')) {
+        this.timelines.find((t) => t.type === 'screenshots-manual').data.push(data)
+      } else {
+        this.timelines.push({
+          data: [data],
+          id: crypto.randomUUID(),
+          name: 'Manual Screenshots',
+          type: 'screenshots-manual'
+        })
+      }
+    },
+    onScreenshotsGeneration(data) {
+      data.forEach((screenshot) => {
+        screenshot.id = crypto.randomUUID()
+      })
+      this.timelines.push({
+        data,
+        id: crypto.randomUUID(),
+        name: 'Screenshots',
+        type: 'screenshots'
+      })
+    },
+    onShotBoundaryDetection(data) {
+      this.timelines.push({
+        data: data.map((shot) => ({ end: shot[1], id: crypto.randomUUID(), start: shot[0] })),
+        id: crypto.randomUUID(),
+        name: 'Shots',
+        type: 'shots'
+      })
     },
     redo() {
       this.$patch(useUndoStore().redo('undoable'))
+    },
+    renameTimeline(id, name) {
+      this.getTimelineById(id).name = name
+    },
+    runShotBoundaryDetection() {
+      api.runShotBoundaryDetection(useMainStore().video)
+    },
+    splitSegment(timelineId, segmentId, position) {
+      const timeline = this.getTimelineById(timelineId)
+      const index = timeline.data.findIndex((s) => s.id === segmentId)
+      const segment = timeline.data[index]
+      timeline.data.splice(index + 1, 0, {
+        end: segment.end,
+        id: crypto.randomUUID(),
+        start: position
+      })
+      segment.end = position - 1
+    },
+    undo() {
+      this.$patch(useUndoStore().undo('undoable'))
     }
   }
 })
