@@ -1,32 +1,32 @@
-import { resolve } from 'path'
-import { defineConfig, mergeConfig } from 'vite'
-import vue from '@vitejs/plugin-vue'
 import Vuetify, { transformAssetUrls } from 'vite-plugin-vuetify'
+import { defineConfig, mergeConfig } from 'vite'
 import VueRouter from 'unplugin-vue-router/vite'
+import { resolve } from 'path'
+import vue from '@vitejs/plugin-vue'
 
 const baseConfig = {
-  root: './src/renderer',
   define: {
     isElectron: false
-  },
-  resolve: {
-    alias: {
-      '@renderer': resolve('src/renderer/src')
-    }
   },
   plugins: [
     VueRouter({ root: process.cwd() + '/src/renderer' }),
     vue({ template: { transformAssetUrls } }),
     Vuetify()
-  ]
+  ],
+  resolve: {
+    alias: {
+      '@renderer': resolve('src/renderer/src')
+    }
+  },
+  root: './src/renderer'
 }
 
 const developmentConfig = {
   server: {
     proxy: {
       '/api': {
-        target: 'http://localhost:8000',
-        changeOrigin: true
+        changeOrigin: true,
+        target: 'http://localhost:8000'
       }
     }
   }
@@ -35,7 +35,6 @@ const developmentConfig = {
 export default defineConfig(({ mode }) => {
   if (mode === 'development') {
     return mergeConfig(baseConfig, developmentConfig)
-  } else {
-    return baseConfig
   }
+  return baseConfig
 })
